Migrate vis_type_metric to kibana/new platform (#63096)
* Move vis_type_metric to Kibana Platform

* Adapt i18n

* Cleanup tests

* Modify CODEOWNERS

* Cleanup SCSS
