+++ .i18nrc.json
@@ -47,7 +47,7 @@
     "uiActions": "src/plugins/ui_actions",
     "visDefaultEditor": "src/plugins/vis_default_editor",
     "visTypeMarkdown": "src/legacy/core_plugins/vis_type_markdown",
-    "visTypeMetric": "src/legacy/core_plugins/vis_type_metric",
+    "visTypeMetric": "src/plugins/vis_type_metric",
     "visTypeTable": "src/legacy/core_plugins/vis_type_table",
     "visTypeTagCloud": "src/legacy/core_plugins/vis_type_tagcloud",
     "visTypeTimeseries": ["src/legacy/core_plugins/vis_type_timeseries", "src/plugins/vis_type_timeseries"],
