--- src/legacy/core_plugins/vis_type_metric/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "metric_vis",
-  "version": "kibana"
-}
