+++ src/plugins/vis_type_metric/kibana.json
@@ -1,0 +1,8 @@
+{
+  "id": "visTypeMetric",
+  "version": "8.0.0",
+  "kibanaVersion": "kibana",
+  "server": true,
+  "ui": true,
+  "requiredPlugins": ["data", "visualizations", "charts","expressions"]
+}
