+++ src/plugins/vis_type_metric/public/_metric_vis.scss
