+++ src/plugins/vis_type_metric/public/components/metric_vis_component.tsx
@@ -22,12 +22,12 @@ import React, { Component } from 'react';
 import { isColorDark } from '@elastic/eui';
 import { MetricVisValue } from './metric_vis_value';
 import { Input } from '../metric_vis_fn';
-import { FieldFormatsContentType, IFieldFormat } from '../../../../../plugins/data/public';
-import { KibanaDatatable } from '../../../../../plugins/expressions/public';
-import { getHeatmapColors } from '../../../../../plugins/charts/public';
+import { FieldFormatsContentType, IFieldFormat } from '../../../data/public';
+import { KibanaDatatable } from '../../../expressions/public';
+import { getHeatmapColors } from '../../../charts/public';
 import { VisParams, MetricVisMetric } from '../types';
 import { getFormatService } from '../services';
-import { SchemaConfig, ExprVis } from '../../../../../plugins/visualizations/public';
+import { SchemaConfig, ExprVis } from '../../../visualizations/public';
 
 export interface MetricVisComponentProps {
   visParams: VisParams;
