+++ src/plugins/vis_type_metric/public/components/metric_vis_value.test.tsx
