+++ src/plugins/vis_type_metric/public/metric_vis_type.test.ts
@@ -17,9 +17,13 @@
  * under the License.
  */
 
-import { PluginInitializerContext } from '../../../../core/public';
-import { MetricVisPlugin as Plugin } from './plugin';
+import { createMetricVisTypeDefinition } from './metric_vis_type';
+import { MetricVisComponent } from './components/metric_vis_component';
 
-export function plugin(initializerContext: PluginInitializerContext) {
-  return new Plugin(initializerContext);
-}
+describe('metric_vis - createMetricVisTypeDefinition', () => {
+  it('has metric vis component set', () => {
+    const def = createMetricVisTypeDefinition();
+
+    expect(def.visConfig.component).toBe(MetricVisComponent);
+  });
+});
